requirements.txt
matplotlib
numpy
scipy
soundfile
pyroomacoustics
pyyaml
torch
torchvision
torchaudio
sounddevice
librosa
pysndfx
tqdm
argparse
sox
wandb